+++ blocklets/ai-runtime/api/src/routes/image.ts
@@ -2,7 +2,7 @@ import { rmSync } from 'fs';
 
 import { call, getComponentMountPoint } from '@blocklet/sdk/lib/component';
 import config from '@blocklet/sdk/lib/config';
-import { Router } from 'express';
+import { NextFunction, Request, Response, Router } from 'express';
 import mime from 'mime-types';
 import multer from 'multer';
 import { nanoid } from 'nanoid';
@@ -33,36 +33,55 @@ const upload = multer({
   },
 });
 
-router.post('/upload', upload.array('images', 3), async (req, res) => {
-  const files = req.files as Express.Multer.File[];
-  if (!files || files.length === 0) {
-    res.status(400).json({ error: 'Please select an image to upload' });
-    return;
-  }
+router.post(
+  '/upload',
+  (req: Request & { isSingleUpload?: boolean }, res: Response, next: NextFunction) => {
+    const isSingleUpload = req.header('X-Single-Upload') === 'true';
+    const uploadMiddleware = isSingleUpload ? upload.single('image') : upload.array('images', 3);
 
-  try {
-    const uploadPromises = files.map(async (file) => {
-      const { data: result } = await call<{ filename: string }>({
-        name: 'image-bin',
-        path: '/api/sdk/uploads',
-        headers: { 'x-user-did': config.env.appId },
-        data: { type: 'path', data: file.path, filename: file.originalname },
+    uploadMiddleware(req, res, (err) => {
+      if (err) {
+        res.status(400).json({ error: err.message });
+        return;
+      }
+
+      req.isSingleUpload = isSingleUpload;
+
+      next();
+    });
+  },
+  async (req: Request & { isSingleUpload?: boolean }, res: Response) => {
+    const files = (req.isSingleUpload ? [req.file] : req.files) as Express.Multer.File[];
+
+    if (!files || files.length === 0) {
+      res.status(400).json({ error: 'Please select an image to upload' });
+      return;
+    }
+
+    try {
+      const uploadPromises = files.map(async (file) => {
+        const { data: result } = await call<{ filename: string }>({
+          name: 'image-bin',
+          path: '/api/sdk/uploads',
+          headers: { 'x-user-did': config.env.appId },
+          data: { type: 'path', data: file.path, filename: file.originalname },
+        });
+
+        return { url: joinURL(config.env.appUrl, getComponentMountPoint('image-bin'), 'uploads', result.filename) };
       });
 
-      return { url: joinURL(config.env.appUrl, getComponentMountPoint('image-bin'), 'uploads', result.filename) };
-    });
-
-    const results = await Promise.all(uploadPromises);
-    res.json({ uploads: results });
-  } finally {
-    files.forEach((file) => {
-      try {
-        rmSync(file.path, { recursive: true, force: true });
-      } catch (err) {
-        console.error(`Failed to remove file ${file.path}:`, err);
-      }
-    });
+      const results = await Promise.all(uploadPromises);
+      res.json({ uploads: req.isSingleUpload ? results[0] : results });
+    } finally {
+      files.forEach((file) => {
+        try {
+          rmSync(file.path, { recursive: true, force: true });
+        } catch (err) {
+          console.error(`Failed to remove file ${file.path}:`, err);
+        }
+      });
+    }
   }
-});
+);
 
 export default router;