+++ packages/ai-runtime/src/front/runtime-components/SimpleChat/OpeningMessageView.tsx
@@ -10,12 +10,15 @@ import { useComponentPreferences } from '../../contexts/ComponentPreferences';
 import { useCurrentAgent } from '../../contexts/CurrentAgent';
 import { CurrentMessageOutputProvider } from '../../contexts/CurrentMessage';
 import { useEntryAgent } from '../../contexts/EntryAgent';
+import { useSession } from '../../contexts/Session';
 import { useOpeningMessage, useOpeningQuestions, useProfile } from '../../hooks/use-appearances';
 import { MessageBodyContainer } from './MessageView';
 import type { SimpleChatPreferences } from '.';
 
-const OpeningMessageView = memo(({ isMessagesEmpty }: { isMessagesEmpty?: boolean }) => {
+const OpeningMessageView = memo(() => {
   const { aid } = useEntryAgent();
+
+  const isMessagesEmpty = useSession((s) => !s.messages?.length);
 
   const { hideAgentAvatar, backgroundImage } = useComponentPreferences<SimpleChatPreferences>() ?? {};
   const hasBg = !!backgroundImage?.url;
@@ -49,14 +52,14 @@ const OpeningMessageView = memo(({ isMessagesEmpty }: { isMessagesEmpty?: boolea
 
   const children = (
     <MessageBodyContainer>
-      {openingMessage && openingMessageOutput && openingMessageComponentId && (
+      {openingMessage?.message && openingMessageOutput && openingMessageComponentId && (
         <CurrentMessageOutputProvider output={openingMessageOutput} outputValue={undefined}>
           <CustomComponentRenderer
             aid={aid}
             output={{ id: openingMessageOutput.id }}
-            instanceId={openingMessageOutput.id}
+            instanceId={`${agent.id}-${openingMessageOutput.id}`}
             componentId={openingMessageComponentId}
-            properties={openingMessageOutput.appearance?.componentProperties}
+            properties={openingMessageOutput?.appearance?.componentProperties}
           />
         </CurrentMessageOutputProvider>
       )}
@@ -66,7 +69,7 @@ const OpeningMessageView = memo(({ isMessagesEmpty }: { isMessagesEmpty?: boolea
           <CustomComponentRenderer
             aid={aid}
             output={{ id: openingQuestionsOutput.id }}
-            instanceId={agent.id}
+            instanceId={`${agent.id}-${openingQuestionsOutput.id}`}
             componentId={openingQuestionsComponentId}
             properties={openingQuestionsOutput?.appearance?.componentProperties}
           />