+++ packages/ai-runtime/src/types/assistant/index.ts
@@ -347,7 +347,7 @@ export interface ParameterBase {
   placeholder?: string;
   helper?: string;
   required?: boolean;
-  from?: 'editor' | 'agentParameter' | 'knowledgeParameter' | 'blockletAPIParameter';
+  from?: 'editor' | 'agentParameter' | 'knowledgeParameter' | 'blockletAPIParameter' | 'imageBlenderParameter';
   hidden?: boolean;
 }
 
@@ -386,6 +386,7 @@ export interface StringParameter extends ParameterBase {
   value?: string;
   defaultValue?: string;
   multiline?: boolean;
+  image?: boolean;
   minLength?: number;
   maxLength?: number;
 }